# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "FlashcardSuggestion": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "region_id": {
            "title": "Region Id",
            "type": "string"
          },
          "source_slide": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Slide"
          },
          "status": {
            "enum": ["suggested", "approved", "rejected"],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "region_id",
          "question",
          "answer",
          "source_slide",
          "status"
        ],
        "title": "FlashcardSuggestion",
        "type": "object"
      },
      "GraphEdge": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": ["id", "source", "target", "label"],
        "title": "GraphEdge",
        "type": "object"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["id", "label", "type"],
        "title": "GraphNode",
        "type": "object"
      },
      "GraphResponse": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": ["nodes", "edges"],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchResponse": {
        "properties": {
          "highlights": {
            "items": {
              "$ref": "#/components/schemas/TextSpan"
            },
            "title": "Highlights",
            "type": "array"
          },
          "passage": {
            "title": "Passage",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "region_id": {
            "title": "Region Id",
            "type": "string"
          },
          "similarity_score": {
            "title": "Similarity Score",
            "type": "number"
          },
          "slide_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Number"
          },
          "status": {
            "enum": ["matched", "uncertain", "no_match"],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "region_id",
          "status",
          "slide_number",
          "passage",
          "highlights",
          "similarity_score",
          "reason"
        ],
        "title": "MatchResponse",
        "type": "object"
      },
...
      "TextSpan": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": ["text", "x", "y", "width", "height"],
        "title": "TextSpan",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API for turning notebook pages into interactive study surfaces.",
    "title": "Traceback API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions/{session_id}/flashcards/generate": {
      "post": {
        "operationId": "generate_flashcards_api_sessions__session_id__flashcards_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlashcardSuggestion"
                  },
                  "title": "Response Generate Flashcards Api Sessions  Session Id  Flashcards Generate Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Flashcards",
        "tags": ["learning"]
      }
    },
    "/api/sessions/{session_id}/graph": {
      "get": {
        "operationId": "get_graph_api_sessions__session_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph",
        "tags": ["learning"]
      }
    },
...
    "/api/sessions/{session_id}/regions/{region_id}/match": {
      "post": {
        "operationId": "match_session_region_api_sessions__session_id__regions__region_id__match_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "region_id",
            "required": true,
            "schema": {
              "title": "Region Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Match Session Region",
        "tags": ["matches"]
      }
    }
  }
}